♻️ refact : 이모티콘 핸들러  리팩토링
